Move IndexingProgressListener to beter package
