+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/gui/dialog/progress/fileindexer/IndexingProgressDialog.java
@@ -5,6 +5,7 @@ import org.lodder.subtools.multisubdownloader.actions.ActionException;
 import org.lodder.subtools.multisubdownloader.gui.dialog.Cancelable;
 import org.lodder.subtools.multisubdownloader.gui.dialog.ProgressDialog;
 import org.lodder.subtools.multisubdownloader.gui.extra.progress.StatusMessenger;
+import org.lodder.subtools.multisubdownloader.listeners.IndexingProgressListener;
 
 public class IndexingProgressDialog extends ProgressDialog implements IndexingProgressListener {
