+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/listeners/IndexingProgressListener.java
@@ -1,4 +1,4 @@
-package org.lodder.subtools.multisubdownloader.gui.dialog.progress.fileindexer;
+package org.lodder.subtools.multisubdownloader.listeners;
 
 import org.lodder.subtools.multisubdownloader.gui.dialog.progress.StatusListener;
